requirements.txt
discord
python-dotenv
bs4
requests
spotipy
